Move test datastores to test. env package
